+++ routes/api-root.js
@@ -1,7 +1,7 @@
 import express from 'express';
 const router = express.Router();
 import {router as s} from './s';
-import {auth} from "./auth/auth";
+import {authRouter} from "./auth/auth";
 
 router.get('/', (req, res, next) => {
   console.log('api-root');
@@ -12,7 +12,7 @@ router.get('/', (req, res, next) => {
   res.json(message);
 });
 
-router.use('/auth', auth);
+router.use('/auth', authRouter);
 
 router.use('/s', s);
