+++ routes/auth/helpers.js
@@ -1,4 +1,5 @@
 import {studentLogin} from "../../resolvers/auth/student-login";
+import {createStudent} from "../../resolvers/auth/create-student";
 
 /**
  * Extracts login params from req body.
@@ -29,12 +30,12 @@ export const extractAccountData = (body) => {
       type: body.type
     }
   } catch (e) {
-      return null;
+    return null;
   }
 };
 
 export const createAccountMap = {
-  student: () => null,
+  student: () => createStudent,
   teacher: () => null,
 };
 
@@ -42,22 +43,3 @@ export const loginMap = {
   student: studentLogin,
   teacher: () => console.log('teacher')
 };
-
-
-import bcrypt from 'bcrypt';
-
-export async function verifyPwd(hash, pwd) {
-  return await bcrypt.compare(pwd, hash);
-}
-
-
-export function buildTokenPayload(user) {
-  delete user.pwd;
-  return user;
-}
-
-
-export async function hashPwd(pwd) {
-  const rounds = 12;
-  return await bcrypt.hash(pwd, rounds);
-}
